[SWM-161] refact: apply Transactional(readOnly=true) as default and remove unusing method

+++ src/main/java/com/climbx/climbx/user/dto/DailyHistoryResponseDto.java
@@ -10,7 +10,7 @@ public record DailyHistoryResponseDto(
     Integer value
 ) {
 
-    public DailyHistoryResponseDto(java.sql.Date date, Integer value) {
-        this(date.toLocalDate(), value);
+    public DailyHistoryResponseDto(java.sql.Date date, Long value) {
+        this(date.toLocalDate(), value.intValue());
     }
 }
